ai/playground/image-description-benchmark/pyproject.toml
[project]
name = "image-description-benchmark"
version = "0.1.0"
description = "Benchmark OpenAI LLMs on image description tasks"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.9.0",
    "google-genai>=1.38.0",
    "openai>=1.108.1",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "rfdetr>=1.2.1",
    "supervision>=0.26.1",
    "tabulate>=0.9.0",
    "websocket>=0.2.1",
    "websocket-client>=1.8.0",
    "websockets>=15.0.1",
]

[project.scripts]
benchmark = "image_description_benchmark.main:main"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "ruff>=0.13.1",
]
